requirements.txt
pytorch-lightning==1.0.2
GitPython==3.1.12
pandas==1.3.4
scikit-learn==0.24.2
pyarrow==5.0.0
pip==22.0.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
comet-ml==3.2.10
